+++ cyclus/pyagents.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals, print_function
 
 from cyclus.agents import Region, Institution, Facility
 from cyclus import typesystem as ts
-from cyclus.lib import cy_large_double
+from cyclus.lib import CY_LARGE_DOUBLE
 
 class NullRegion(Region):
     """A simple do nothing region."""
@@ -32,7 +32,7 @@ class Sink(Facility):
         uitype="recipe",
         )
     max_inv_size = ts.Double(
-        default=cy_large_double(),
+        default=CY_LARGE_DOUBLE(),
         doc="total maximum inventory size of sink facility",
         uilabel= "Maximum Inventory",
         tooltip="sink maximum inventory size",